1.0.47: remove meta.embedded.slice.fish
It's not really a sub-language.

Really, what happens between the [ ] is the same as what happens outside of them: `$(subcommand)`` or `(subcommand)` are expended depending on if the list variable is already in a string.

`..` is not a low-binding keyword. What happens is: the string passed into the [ ] is interpreted by an internal fish function, so something like `$list_variable[$(echo '1..2')]` is valid and the same as `$list_variable[1..2]`. To bollow from Python, it's more like a custom `__getitem__` implementation that has its own meaning of `..`, than a syntax construct.

However, the fish shell itself highlights the stuff between [ ] differently, so maybe I will do this one day. A meta language that highlights everything inside as a variable. If coming from a double quoted string, include only variables, if coming from something else, include all usual language constructs.