remove .. from slice syntax
`..` is not a low-binding keyword. What happens is: the string passed into the [ ] is interpreted by an internal fish function, so something like `$list_variable[$(echo '1..2')]` is valid and the same as `$list_variable[1..2]`. To bollow from Python, it's more like a custom `__getitem__` implementation that has its own meaning of `..`, than a syntax construct.