1.0.48: fix "$argv[(cmd)]" from incorrectly showing expansion
What happens between the [ ] is the same as what happens outside of them: `$(subcommand)` or `(subcommand)` are expended depending on if the list variable is already in a string. Implement this by adding more special cases - when a subshell occurs inside of a string, and when it doesn't.

Make it so the text inside the [ ] is highlighted as a variable. It is useful to know that the when you write `$argv[1..2]`, `$argv` is taking over `[1..2]` and it's not just the next string. The fish shell itself highlights this differently, so we should too.

With this, all that the slice meta.embedded language does is allow `[ ]` to be bracket-pair matched, and highlight what's in the brackets as a variable.

Originally I made a branch giving up on slices entirely: `remove-slice`, commit 28a9363b3c418384ca4480dac4e2d2dda53f8c6b.