remove CONTRIBUTING.md
Written by a contributor, but with instructions very specific to that one contribution.
I don't really understand these instrunctions, so stop telling it to "everyone".

I've had only a few contributions over the years so it's not like I need to put out a memo to a huge number of contributors to shape their contributions.