prettier: format fish.tmLanguage.json, switch pre-commit hook to simple-git-hooks
husky was apparently not working, since this fish.tmLanguage.json had drifted out of format.

I attempted repair my husky config, update husky, etc. but it was rather complicated with some setup script I had to run outside of `yarn prepare`, some .husky/ directory of binaries, and finally when it did run it inexplicably tried to run `yarn test` pre-commit which I never told it to do.

Switch to simple-git-hooks, which worked on the first try. This is just the lifecycle of software; you start with something simple built for an unaddressed use case. Then you add more features as use cases change. Finally, a new software is created that better addersses where the use case is today while not supporting the old use case, which takes your place.